+++ src/app/app-extras.module.ts
@@ -4,12 +4,17 @@ import { SkyDemoModule } from '../demo';
 
 require('style-loader!@skyux/theme/css/sky.css');
 
+import {
+  AppSkyModule
+} from './app-sky.module';
+
 @NgModule({
   entryComponents: [],
   imports: [
     SkyDemoModule
   ],
   exports: [
+    AppSkyModule,
     SkyDemoModule
   ],
   providers: []