+++ src/app/app-sky.module.ts
@@ -1,0 +1,195 @@
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  SkyAvatarModule
+} from '@skyux/avatar';
+
+import {
+  SkyColorpickerModule
+} from '@skyux/colorpicker';
+
+import {
+  SkyNumericModule,
+  SkyMediaQueryModule
+} from '@skyux/core';
+
+import {
+  SkyTimepickerModule,
+  SkyDatepickerModule
+} from '@skyux/datetime';
+
+import {
+  SkyErrorModule
+} from '@skyux/errors';
+
+import {
+  SkyFlyoutModule
+} from '@skyux/flyout';
+
+import {
+  SkyCheckboxModule,
+  SkyRadioModule,
+  SkyFileAttachmentsModule
+} from '@skyux/forms';
+
+import {
+  SkyGridModule
+} from '@skyux/grids';
+
+import {
+  SkyWaitModule,
+  SkyIconModule,
+  SkyTokensModule,
+  SkyTextHighlightModule,
+  SkyAlertModule,
+  SkyLabelModule,
+  SkyKeyInfoModule,
+  SkyHelpInlineModule
+} from '@skyux/indicators';
+
+import {
+  SkyToolbarModule,
+  SkyTextExpandModule,
+  SkyTextExpandRepeaterModule,
+  SkyFluidGridModule,
+  SkyPageSummaryModule,
+  SkyDefinitionListModule,
+  SkyActionButtonModule,
+  SkyCardModule
+} from '@skyux/layout';
+
+import {
+  SkyListModule,
+  SkyListToolbarModule,
+  SkyListFiltersModule,
+  SkyListSecondaryActionsModule,
+  SkyListPagingModule
+} from '@skyux/list-builder';
+
+import {
+  SkyListViewChecklistModule
+} from '@skyux/list-builder-view-checklist';
+
+import {
+  SkyListViewGridModule
+} from '@skyux/list-builder-view-grids';
+
+import {
+  SkySortModule,
+  SkyRepeaterModule,
+  SkyPagingModule,
+  SkyFilterModule,
+  SkyInfiniteScrollModule
+} from '@skyux/lists';
+
+import {
+  SkySearchModule,
+  SkyLookupModule,
+  SkyAutocompleteModule
+} from '@skyux/lookup';
+
+import {
+  SkyModalModule,
+  SkyConfirmModule
+} from '@skyux/modals';
+
+import {
+  SkyNavbarModule
+} from '@skyux/navbar';
+
+import {
+  SkyPopoverModule,
+  SkyDropdownModule
+} from '@skyux/popovers';
+
+import {
+  SkyProgressIndicatorModule
+} from '@skyux/progress-indicator';
+
+import {
+  SkySelectFieldModule
+} from '@skyux/select-field';
+
+import {
+  SkyVerticalTabsetModule,
+  SkyTabsModule,
+  SkySectionedFormModule
+} from '@skyux/tabs';
+
+import {
+  SkyTilesModule
+} from '@skyux/tiles';
+
+import {
+  SkyToastModule
+} from '@skyux/toast';
+
+import {
+  SkyUrlValidationModule,
+  SkyEmailValidationModule
+} from '@skyux/validation';
+
+@NgModule({
+  exports: [
+    SkyActionButtonModule,
+    SkyAlertModule,
+    SkyAutocompleteModule,
+    SkyAvatarModule,
+    SkyCardModule,
+    SkyCheckboxModule,
+    SkyColorpickerModule,
+    SkyConfirmModule,
+    SkyDatepickerModule,
+    SkyDefinitionListModule,
+    SkyDropdownModule,
+    SkyEmailValidationModule,
+    SkyErrorModule,
+    SkyFileAttachmentsModule,
+    SkyFilterModule,
+    SkyFluidGridModule,
+    SkyFlyoutModule,
+    SkyGridModule,
+    SkyHelpInlineModule,
+    SkyIconModule,
+    SkyInfiniteScrollModule,
+    SkyKeyInfoModule,
+    SkyLabelModule,
+    SkyListFiltersModule,
+    SkyListModule,
+    SkyListPagingModule,
+    SkyListSecondaryActionsModule,
+    SkyListToolbarModule,
+    SkyListViewChecklistModule,
+    SkyListViewGridModule,
+    SkyLookupModule,
+    SkyMediaQueryModule,
+    SkyModalModule,
+    SkyNavbarModule,
+    SkyNumericModule,
+    SkyPageSummaryModule,
+    SkyPagingModule,
+    SkyPopoverModule,
+    SkyProgressIndicatorModule,
+    SkyRadioModule,
+    SkyRepeaterModule,
+    SkySearchModule,
+    SkySectionedFormModule,
+    SkySelectFieldModule,
+    SkySortModule,
+    SkyTabsModule,
+    SkyTextExpandModule,
+    SkyTextExpandRepeaterModule,
+    SkyTextHighlightModule,
+    SkyTilesModule,
+    SkyTimepickerModule,
+    SkyToastModule,
+    SkyTokensModule,
+    SkyToolbarModule,
+    SkyUrlValidationModule,
+    SkyVerticalTabsetModule,
+    SkyWaitModule
+  ]
+})
+export class AppSkyModule { }