+++ src/demos/demo.service.ts
@@ -775,6 +775,15 @@ export class SkyDemoService {
         {
           name: 'modal-demo-context.ts',
           fileContents: require('!!raw-loader!./modal/modal-demo-context.ts')
+        },
+        {
+          name: 'modal-demo-close-confirmation.component.ts',
+          fileContents: require('!!raw-loader!./modal/modal-demo-close-confirmation.component.ts'),
+          componentName: 'SkyModalDemoCloseConfirmationComponent'
+        },
+        {
+          name: 'modal-demo-close-confirmation.component.html',
+          fileContents: require('!!raw-loader!./modal/modal-demo-close-confirmation.component.html')
         }
       ]
     },
@@ -1283,7 +1292,7 @@ export class SkyDemoService {
 
   public getComponent(name: string): any {
     const found = this.components.find((component: any) => {
-      return component.component.name === name;
+      return component.name === name;
     });
 
     return found;