+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 - Initial major release
 
+# 1.0.0-alpha.5 (2019-06-13)
+
+- Added an instance of the `sky-repeater-item-context-menu` component to the repeater component demo. [#3](https://github.com/blackbaud/skyux2-demos/pull/3)
+
+# 1.0.0-alpha.4 (2019-06-06)
+
+- Added support for `@skyux-sdk/builder`.
+
+# 1.0.0-alpha.3 (2019-06-05)
+
+- Added `.sky-margin-inline-default` to all demo button elements.
+
+# 1.0.0-alpha.2 (2019-06-05)
+
+- Removed broken `*ngIf` attributes from repeater component demo.
+
+# 1.0.0-alpha.1 (2019-06-05)
+
+- Added `SkyModalDemoCloseConfirmationComponent` to demo service's `components` array.
+
 # 1.0.0-alpha.0 (2019-06-04)
 
 - Initial alpha release.